Prevent too fast VRAM access on MSX1
Thanks to Grauw for the review and tips!

Before this fix, VDP VRAM access was too fast, which was at least
visible in the C-BIOS MSX1 logo, but probably also on many games using
C-BIOS MSX1. This wasn't an issue before openMSX implemented accurate
VDP VRAM access timing, but now it clearly showed...