Replaced "find" by "cp -R" in "make dist"
The "find" based approach was necessary to exclude the ".svn" dirs
back when we were using Subversion. However, "cp --parents" is a
GNU-ism which breaks on Mac OS X (and on *BSD, I assume). So now we
use just "cp -R" to copy whole directories, which is portable and
in my test didn't copy any junk files.