Final "make dist" portability improvements
There is no "sha1sum" on Mac OS X, but "shasum -a1" works on both Linux
and OS X.

The syntax for "mktemp" is different on Mac OS X and Linux. There is no
real need though to put this temporary file in a dedicated temp dir;
instead we can just put it in derived/dist instead. This has the added
advantage that if the commands using the temporary file break, the
leftover file is easier to find when trying to figure out the problem.

Removed "-s" option to "sed": it doesn't seem to be doing anything in
this particular use case and it doesn't exist on BSD sed. Maybe it was
a leftover from a different way of using "sed".

Pass non-empty suffix to "sed -i": GNU sed requires the suffix to
immediately follow the "-i", but BSD sed will not recognize an empty
suffix it if immediately follows "-i". So instead we do create backup
files and remove them later.